Reduce the verbosity of Geant4 Physics lists printouts
By default, the G4 physics list FTFP_BERT (that is usually used) will print a fairly long printout that is not captured by framework logging. It looks something like this:

```
 hInelastic FTFP_BERT : threshold between BERT and FTFP is over the interval
 for pions :   3 to 6 GeV
 for kaons :   3 to 6 GeV
 for proton :  3 to 6 GeV
 for neutron : 3 to 6 GeV

=======================================================================
======                 Electromagnetic Physics Parameters      ========
=======================================================================
LPM effect enabled                                 1
Enable creation and use of sampling tables         0
Apply cuts on all EM processes                     0
Use general process                                0
...
```

This info may be good to have, but is the same for (unchanged) physics list of the same geant4 versions. For better readability of the logs, we make this optional, with default off.